+++ pyproject.toml
@@ -25,7 +25,7 @@ scipy = "1.11.3"
 seaborn = "0.13.2"
 snakemake = "8.10.0"
 tqdm = "4.66.5"
-dscript = "0.2.6"
+dscript = "0.2.8"
 fastdsd = "1.0.1"
 py4cytoscape = "1.9.0"
 recipe-cluster = "0.0.2"
